refactor: move study-page.ts into app/study/ subfolder
