+++ src/vite/src/entry-study.ts
@@ -10,7 +10,7 @@ import "./assets/css/study-base.build.css";
 
 import { domGetById, domContentLoaded } from "@lib/dom";
 import VueInstance from "@lib/helpers/vue-instance";
-import StudyPage from "@app/study-page";
+import StudyPage from "@/app/study/study-page";
 import MyStoriesTable from "@app/study/MyStoriesTable.vue";
 
 domContentLoaded(() => {
